Feature agentsmith

+++ lib/point.go
@@ -1,6 +1,6 @@
 package lib
 
 type Point struct {
-	X int64 `json:"x"`
-	Y int64 `json:"y"`
+	X int `json:"x"`
+	Y int `json:"y"`
 }
